# crates/clawft-services/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeftOS gateway API",
    "version": "0.8.1",
    "description": "REST, WebSocket and MCP surface of `weft gateway`. Every route except GET /api/health needs a bearer token issued by the kernel daemon (`weft token issue`); there is no mint route. Request and response bodies are mostly untyped here; see docs/ui/api-reference.md. Hand-maintained: tests/openapi.rs fails when a served route is missing or stale."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/agents": {
      "get": {
        "summary": "List agents",
        "operationId": "get_api_agents",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        }
      }
    },
...
    "/api/agents/{name}": {
      "get": {
        "summary": "Get agent",
        "operationId": "get_api_agents_name_",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "summary": "Delete agent by pid",
        "operationId": "delete_api_agents_name_",
        "tags": [
          "handlers"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/auth/revoke": {
      "post": {
        "summary": "Revoke the caller's own token",
        "operationId": "post_api_auth_revoke",
        "tags": [
          "handlers"
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "description": "Forwards the authenticating token's own id to the daemon. Cannot name another token."
      }
    },
...
    "/custody/witness": {
      "post": {
        "summary": "Custody witness",
        "operationId": "post_custody_witness",
        "tags": [
          "http_facade"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        }
      }
    },
    "/events": {
      "get": {
        "summary": "Server-sent kernel events",
        "operationId": "get_events",
        "tags": [
          "http_facade"
        ],
        "responses": {
          "200": {
            "description": "text/event-stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        }
      }
    },
    "/mcp": {
      "post": {
        "summary": "MCP JSON-RPC (full profile)",
        "operationId": "post_mcp",
        "tags": [
          "mcp"
        ],
        "responses": {
          "200": {
            "description": "JSON-RPC response"
          },
          "202": {
            "description": "Notification accepted"
          },
          "400": {
            "description": "Malformed JSON"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        },
        "description": "initialize, tools/list, tools/call and notifications. Only mounted when the gateway has an MCP shell. Owner-equivalent: every tool is reachable.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonRpcRequest"
              }
            }
          }
        }
      }
    },
    "/ws": {
      "get": {
        "summary": "WebSocket upgrade (token via Authorization header or ?token=)",
        "operationId": "get_ws",
        "tags": [
          "ws"
        ],
        "responses": {
          "101": {
            "description": "Switching protocols"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        }
      }
    },
    "/api/openapi.json": {
      "get": {
        "summary": "This document",
        "operationId": "get_api_openapi_json",
        "tags": [
          "openapi"
        ],
        "responses": {
          "200": {
            "description": "OpenAPI 3.1 document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/DaemonUnavailable"
          }
        }
      }
    },
...
    "/api/console/token": {
      "post": {
        "summary": "Mint a short-lived read-only token for one project from a tailnet identity",
        "operationId": "post_api_console_token",
        "tags": [
          "console"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "project"
                ],
                "properties": {
                  "project": {
                    "type": "string",
                    "description": "Project ULID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{token, expires_at, project}"
          },
          "400": {
            "description": "Bad body"
          },
          "403": {
            "description": "Not a tailnet peer, login not allowlisted, or cross-origin"
          },
          "404": {
            "description": "Unknown project, or tailnet identity is not enabled"
          },
          "503": {
            "description": "whois or daemon unavailable"
          }
        },
        "description": "Disabled (404) unless gateway.tailnetIdentity.enabled. No bearer: the caller is identified by its TCP peer address and tailscale whois."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Daemon-issued owner token (wft_...). Project tokens are refused."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, unknown, expired, revoked or non-owner token",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "DaemonUnavailable": {
        "description": "The daemon that owns tokens is unreachable",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "remedy": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "JsonRpcRequest": {
        "type": "object",
        "required": [
          "jsonrpc",
          "method"
        ],
        "properties": {
          "jsonrpc": {
            "const": "2.0"
          },
          "id": {},
          "method": {
            "type": "string"
          },
          "params": {}
        }
      }
    }
  }
}